[ok] turn

+++ pgx/backgammon.py
@@ -65,7 +65,7 @@ class BackgammonState:
     played_dice_num: jnp.ndarray = jnp.int16(0)
 
     # 黒なら-1, 白なら1
-    turn: jnp.ndarray = jnp.int16(1)
+    turn: jnp.ndarray = jnp.int8(1)
     """
     合法手
     micro action = 6*src+die
@@ -325,7 +325,7 @@ def _init_turn(dice: jnp.ndarray) -> jnp.ndarray:
     サイコロの目が大きい方が手番.
     """
     diff = dice[1] - dice[0]
-    return (diff > 0) * jnp.int16(1) + (diff <= 0) * jnp.int16(-1)
+    return (diff > 0) * jnp.int8(1) + (diff <= 0) * jnp.int8(-1)
 
 
 def _set_playable_dice(dice: jnp.ndarray) -> jnp.ndarray:
